requirements.txt
djangorestframework==3.15.2
pylint-django==2.5.5
coverage==7.5.4
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-environ>=0.7,<1.0